fixup! feat: autoupdater
